Created the text_resources file + Created a generic translation component + Added this new component in App.js file
