+++ frontend/src/components/eMIB/Background.jsx
@@ -1,5 +1,5 @@
 import React, { Component } from "react";
-import { STRINGS } from "./Emib";
+import STRINGS from "../../text_resources";
 
 class Background extends Component {
   render() {
