+++ frontend/src/text_resources.js
@@ -1,0 +1,83 @@
+import LocalizedStrings from "react-localization";
+
+let STRINGS = new LocalizedStrings({
+  en: {
+    //Home Page
+    testTitle: "eMIB - Sample Test",
+    welcomeMsg: "Welcome to the eMIB Sample Test",
+
+    //HowTo Page
+    howToPageTitle: "HowTo Page",
+    emailInstructions: "Email Instuctions",
+    taskInstructions: "Task Instuctions",
+    howToNextButton: "Start test",
+
+    //Instructions Tab
+    instructionsTabTitle: "Instructions",
+
+    //Background Tab
+    backgroundTabTitle: "Background",
+    backgroundPageTitle: "Background Page",
+    orgChart: "Org Chart",
+    Scenarios: "Scenarios",
+
+    //Inbox Tab
+    inboxTabTitle: "Inbox",
+    inboxPageTitle: "Inbox",
+    taskList: "Tasks List",
+    notePad: "NotePad",
+    textTools: "Text Tools",
+    emailFeatures: "Email Features",
+    progressSaved: "Progress Saved",
+    timer: "Timer",
+
+    //Confirmation Page
+    submissionConfirmed: "Submission Confirmed",
+    exitTest: "Exit Test",
+
+    //Generic
+    nextButton: "Next",
+    submitTestButton: "Submit test"
+  },
+
+  fr: {
+    //Home Page
+    testTitle: "eMIB - Test Pratique",
+    welcomeMsg: "Bienvenu dans le test pratique de eMIB",
+
+    //HowTo Page
+    howToPageTitle: "Page 'How To'",
+    emailInstructions: "Instructions pour les courriel",
+    taskInstructions: "Instuctions pour les tâches",
+    howToNextButton: "Commencer le test",
+
+    //Instructions Tab
+    instructionsTabTitle: "Instructions",
+
+    //Background Tab
+    backgroundTabTitle: "Contexte",
+    backgroundPageTitle: "Page de contexte",
+    orgChart: "Organigramme",
+    Scenarios: "Scénarios",
+
+    //Inbox Tab
+    inboxTabTitle: "Boîte de réception",
+    inboxPageTitle: "Boîte de réception",
+    taskList: "Liste des tâches",
+    notePad: "bloc-notes",
+    textTools: "Outils de texte",
+    emailFeatures: "Fonctions de messagerie",
+    progressSaved: "Progression enregistrée",
+    timer: "Minuteur",
+
+    //Confirmation Page
+    submissionConfirmed: "Soumission Confirmée",
+    exitTest: "Quitter le test",
+
+    //Generic
+    nextButton: "Suivant",
+    submitTestButton: "Soumettre le test"
+  }
+});
+
+export default STRINGS;
